old-tests/submission/W3C/domhtml/HTMLLinkElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>link</title>
<link href="/resources/testharness.css" hreflang="en" media="screen" rev="stylesheet" type="text/css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-link-element.html#dom-link-media" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The media attribute specifies the targeted media.     

    Retrieve the media attribute and examine its value.</p>

<script>
var domtest = async_test("The media attribute specifies the targeted media.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vmedia;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("link");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      vmedia = testNode.media;

      assert_equals(vmedia,"screen");
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
